修改储存类型为Double

+++ src/main/java/cn/fd/sfaddons/ReachPoint/PWay.java
@@ -1,25 +1,24 @@
 package cn.fd.sfaddons.ReachPoint;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public class PWay {
 
-    public PWay(Action action, BigDecimal points, List<String> conditions) {
+    public PWay(Action action, double points, List<String> conditions) {
         this.action = action;
         this.repointM = points;
         this.conditions = conditions;
     }
 
     private final Action action;
-    private final BigDecimal repointM;
+    private final double repointM;
     private final List<String> conditions;
 
     public Action getAction() {
         return action;
     }
 
-    public BigDecimal getRepointM() {
+    public double getRepointM() {
         return repointM;
     }
